refactor(project): create a directory named old inside landing for last year components
