+++ components/landing/old/About.vue
@@ -41,7 +41,7 @@ export default {
 </script>
 
 <style scoped lang="scss">
-@import '../../assets/mixins.scss';
+@import '../../../assets/mixins';
 div {
   text-align: center;
 }
