+++ components/landing/old/TrophyMobile.vue
