+++ components/landing/old/video.vue
@@ -84,7 +84,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import '../../assets/mixins.scss';
+@import '../../../assets/mixins';
 
 .wrapper {
   display: flex;
