+++ layouts/default.vue
@@ -10,7 +10,7 @@
 
 <script>
 import MyFooter from '../components/MyFooter';
-import Header from '~/components/landing/Header.vue';
+import Header from '~/components/landing/old/Header.vue';
 export default {
   components: { MyFooter, Header },
 };
